% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{add_app}
\alias{add_app}
\title{add App using sbpack}
\usage{
add_app(project, app, app_id, api = "cgc", update_key = FALSE)
}
\arguments{
\item{project}{TThe project ID in the following format: {project_owner}/{project}.}

\item{app}{The app JSON file to upload. Or a \code{cwlProcess} from \code{Rcwl}.}

\item{app_id}{The app ID.}

\item{api}{The cloud api, cgc or cavatica.}

\item{update_key}{Whether to update key.}
}
\description{
add App using sbpack
}
