% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authorization}
\usage{
auth(key, api = "cgc")
}
\arguments{
\item{key}{The Authentication Token of your account from API.}

\item{api}{The cloud api, such as cgc or cavadica.}
}
\description{
Authorization
}
