% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{exec_setting}
\alias{exec_setting}
\title{execution_settings}
\usage{
exec_setting(
  instance_type = "c4.2xlarge",
  max_parallel_instances = 1,
  use_memoization = FALSE
)
}
\arguments{
\item{instance_type}{The instance to run task.}

\item{max_parallel_instances}{Max number of parallel instances.}

\item{use_memoization}{Automatic reuse of precomputed results.}
}
\description{
execution_settings
}
