% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{list_folder}
\alias{list_folder}
\title{list folder}
\usage{
list_folder(id = NULL, project = NULL, recursive = FALSE)
}
\arguments{
\item{id}{The folder ID to list.}

\item{project}{The prject ID.}

\item{recursive}{Whether to list the folder recursively.}
}
\value{
A data.frame of contents in the folder.
}
\description{
list folder
}
